# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MailSessionLanguage = typeof MailSessionLanguage[keyof typeof MailSessionLanguage];


export const MailSessionLanguage = {
  tr: 'tr',
  ru: 'ru',
  en: 'en',
} as const;

export interface MailSession {
  address: string;
  expiresAt: string;
  refreshesUsed: number;
  refreshesLimit: number;
  language: MailSessionLanguage;
}

export interface MailMessage {
  id: string;
  sender: string;
  senderEmail?: string;
  subject: string;
  preview?: string;
  receivedAt: string;
  isRead: boolean;
  /** @nullable */
  verificationCode?: string | null;
  /** @nullable */
  html?: string | null;
  /** @nullable */
  text?: string | null;
}

export interface InboxResponse {
  messages: MailMessage[];
  unreadCount: number;
  checkedAt: string;
}

export type LanguageInputLanguage = typeof LanguageInputLanguage[keyof typeof LanguageInputLanguage];


export const LanguageInputLanguage = {
  tr: 'tr',
  ru: 'ru',
  en: 'en',
} as const;

export interface LanguageInput {
  language: LanguageInputLanguage;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Telegram initData is missing or invalid
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Too many new addresses
 */
export type RateLimitedResponse = ErrorResponse;

/**
 * Telegram Mini App initData sent as Authorization: tma <initData>
 */
export type TelegramInitDataParameter = string;