# lib/api-zod/src/generated/types/inboxResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MailMessage } from './mailMessage';

export interface InboxResponse {
  messages: MailMessage[];
  unreadCount: number;
  checkedAt: Date;
}