# lib/api-zod/src/generated/types/languageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LanguageInputLanguage } from './languageInputLanguage';

export interface LanguageInput {
  language: LanguageInputLanguage;
}