# lib/api-zod/src/generated/types/mailMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MailMessage {
  id: string;
  sender: string;
  senderEmail?: string;
  subject: string;
  preview?: string;
  receivedAt: Date;
  isRead: boolean;
  /** @nullable */
  verificationCode?: string | null;
  /** @nullable */
  html?: string | null;
  /** @nullable */
  text?: string | null;
}