# lib/api-zod/src/generated/types/mailSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MailSessionLanguage } from './mailSessionLanguage';

export interface MailSession {
  address: string;
  expiresAt: Date;
  refreshesUsed: number;
  refreshesLimit: number;
  language: MailSessionLanguage;
}